+++ CHANGES.rst
@@ -9,6 +9,11 @@ improvements
 
 - ``dst_responded`` criterion can be used for SSL measurements too (and is considered in analysis reports too)
 
+fixes
++++++
+
+- cosmetic
+
 0.1.7
 -----